rename chars to field and extract hooks
